# src/main/resources/SwaggerYaml.yml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about
    Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).
    For this sample, you can use the api key `special-key` to test the authorization
    filters."
  version: "1.0.0"
  title: "Swagger Petstore"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
host: "petstore.swagger.io"
basePath: "/v2"
...
schemes:
- "http"
paths:
...
   /Language:
     post:
       tags:
       - "Language"
       summary: "Add a new Language object"
       description: "Add a new Language object value."
       operationId: "addLanguage"
       consumes:
       - "application/json"
       - "application/xml"
       produces:
       - "application/xml"
       - "application/json"
       parameters:
       - in: "body"
         name: "body"
         description: "Pet object that needs to be added to the store"
         required: true
         schema:
           $ref: "#/definitions/Language"
       responses:
         405:
          description: "Invalid input"
   /Language/{ID}:
     get:
       tags:
       - "Language"
       summary: "Find Language by ID"
       description: "Returns a single Language"
       operationId: "getLanguageByID"
       produces:
       - "application/xml"
       - "application/json"
       parameters:
       - name: "ID"
         in: "path"
         description: "ID of Language to return"
         required: true
         type: "integer"
         format: "int64"
       responses:
         200:
           description: "successful operation"
           schema:
             $ref: "#/definitions/Language"
         400:
           description: "Invalid ID supplied"
         404:
           description: "Language not found"
     delete:
       tags:
       - "Language"
       summary: "Deletes a Language"
       description: ""
       operationId: "deleteLanguage"
       produces:
       - "application/xml"
       - "application/json"
       parameters:
       - name: "ID"
         in: "path"
         description: "Pet Language to delete"
         required: true
         type: "integer"
         format: "int64"
       responses:
         400:
           description: "Invalid Language value"
definitions:
...
  Language:
    type: "object"
    required:
    - "ID"
    - "CD"
    - "DESCRIPTION"
    properties:
      ID:
        type: "integer"
        format: "int64"
      CD:
        type: "string"
        format: "int64"
      DESCRIPTION:
        type: "string"
        format: "int64"
    xml:
      name: "Language"